requirements.txt
biopython>=1.81
psycopg2-binary>=2.9.6
pyinstaller>=5.7.0
requests>=2.28.2
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.36.2
keybert>=0.7.0
scikit-learn>=1.3.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
numpy>=1.24.0
hdbscan>=0.8.29
umap-learn>=0.5.3
matplotlib